pyproject.toml
[tool.poetry]
name = "stlit-2"
version = "0.1.0"
description = ""
authors = ["Rajesh J Advani <rajeshjadvani@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.31.0"
python-docx = "^1.1.0"
langchain = "^0.1.6"
chromadb = "^0.4.22"
docx2txt = "^0.8"
nltk = "^3.8.1"
langchain-openai = "^0.0.5"
openai = "^1.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
